requirements.txt
Pillow~=9.5.0
pandas~=2.1.4